BETA: Fixed speed calculation and few bugs with pets: 	* Fixed Speed calculation too much cast to INT in some cases change the stats 	* Fixed admin command for feed the pet send wrong packet and pet don't restore instant 	* Fixed pet speed don't get reduced properly when is hungry 	* Fixed pet Uncontrollable animation don't despair 	* Patch by: Zealar

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/datatables/NpcData.java
@@ -271,16 +271,16 @@ public class NpcData extends DocumentParser
 													{
 														case "walk":
 														{
-															set.set("baseWalkSpd", parseFloat(attrs, "ground"));
-															set.set("baseSwimWalkSpd", parseFloat(attrs, "swim"));
-															set.set("baseFlyWalkSpd", parseFloat(attrs, "fly"));
+															set.set("baseWalkSpd", parseDouble(attrs, "ground"));
+															set.set("baseSwimWalkSpd", parseDouble(attrs, "swim"));
+															set.set("baseFlyWalkSpd", parseDouble(attrs, "fly"));
 															break;
 														}
 														case "run":
 														{
-															set.set("baseRunSpd", parseFloat(attrs, "ground"));
-															set.set("baseSwimRunSpd", parseFloat(attrs, "swim"));
-															set.set("baseFlyRunSpd", parseFloat(attrs, "fly"));
+															set.set("baseRunSpd", parseDouble(attrs, "ground"));
+															set.set("baseSwimRunSpd", parseDouble(attrs, "swim"));
+															set.set("baseFlyRunSpd", parseDouble(attrs, "fly"));
 															break;
 														}
 													}
